Initial commit
A Bluetooth Low Energy (BLE) enabled Harry Potter Sorting Hat.

This project replaces the original electronics in the Harry Potter
Sorting Hat from Barnes and Noble,
https://www.barnesandnoble.com/w/home-gift-harry-potter-sorting-hat/30008992,
with a Nordic nRF5 BLE capable device.

Pins.h contains the definitions for which pins are connected to which
external hardware.

The user can connect to the device with a BLE capable phone using the
Nordic UART serice and send 'g', 'r', 'h', or 's' to select the house
and then a series of random clips from the movies will play, ending
with an announcement of the selected house.

The sound clips are stored on SD card as .wav files and are played back
by pulsing a pin as a 1-bit DAC at 22.05kHz. The mouth is connected to
a hobby servo instead of the original open-loop DC motor and the
position of this servo is based on the volume in upcoming audio
buffers.

The sound clips and the code flow to randomly select between them
originates from https://github.com/gowenrw/arduino_sorting_hat created
by @gowenr, Richard Gowen.

This sample is heavily influenced by Nordic's BLE UART Service
(ble_app_uart) SDK sample.