Generate audio files
Added the https://github.com/gowenrw/arduino_sorting_hat repository
from @gowenrw as a submodule. It contains the required audio clips as
MP3 files. I added an 'audio' rule to my main makefile which is able to
convert these to the required 22.050kHz mono .wav files required by the
playback code that I wrote for the nRF51. The generated files should be
placed in the root folder of the SD card.