Initial commit of iOS application
This is a simple application which connects to the Sorting Hat from an
iPhone for the purpose of selecting the house to be announced. It
automatically reconnects after the hat finishes its announcement and
turns the radio back on (once it no longer needs the CPU resources for
sound playback). The user is presented with 4 buttons, one for each
house, and will need to press the desired button twice to start the
hat announcement so that a single incorrect button tap isn't registered
by mistake.